move chat agents to chat package
