+++ aihub_bot/aihub_bot/bots/chat/openai/StreamOpenaiChatBot.py
@@ -1,13 +1,14 @@
 from botbuilder.core import TurnContext
+from botframework.connector import Channels
 from typing_extensions import override
 
-from aihub_bot.bots.agent.AgentChatBot import AgentChatBot
+from aihub_bot.bots.chat.openai.OpenaiChatBot import OpenaiChatBot
 
 
-class StreamAgentChatBot(AgentChatBot):
+class StreamOpenaiChatBot(OpenaiChatBot):
     @override
     async def on_message_activity(self, turn_context: TurnContext):
-        if turn_context.activity.channel_id == "webchat":
+        if turn_context.activity.channel_id == Channels.webchat:
             return await super().on_message_activity(turn_context)
 
         await self._process_message(turn_context, is_streaming=True)
